Generate sfmt.go and dsfmt.go
sfmt and dsfmt need to build with -msse2 on x86_64 but the flag is not
available on aarch64.  Move the files to *.in as templates and use
xgenflagsfile.bash to generate the final go files from the template
based on whether the build is for x86_64 or some other architecture.

Signed-off-by: Siddhesh Poyarekar <siddhesh@linaro.org>